+++ src/main/java/net/herdspanic/goal/AvoidRainGoal.java
@@ -1,8 +1,10 @@
 package net.herdspanic.goal;
 
+import net.herdspanic.HerdsPanicMain;
 import net.minecraft.entity.ai.NavigationConditions;
 import net.minecraft.entity.ai.goal.AvoidSunlightGoal;
 import net.minecraft.entity.mob.PathAwareEntity;
+import net.minecraft.entity.passive.TameableEntity;
 
 public class AvoidRainGoal extends AvoidSunlightGoal {
     private final PathAwareEntity mob;
@@ -14,7 +16,19 @@ public class AvoidRainGoal extends AvoidSunlightGoal {
 
     @Override
     public boolean canStart() {
-        return this.mob.getWorld().getLevelProperties().isRaining() && NavigationConditions.hasMobNavigation(this.mob);
+        if (!HerdsPanicMain.CONFIG.shelter_seeking) {
+            return false;
+        }
+        if (!NavigationConditions.hasMobNavigation(this.mob)) {
+            return false;
+        }
+        if (!this.mob.getWorld().getLevelProperties().isRaining()) {
+            return false;
+        }
+        if (this.mob instanceof TameableEntity tameableEntity && tameableEntity.isSitting()) {
+            return false;
+        }
+        return true;
     }
 
 }