+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ### Fixed:
 - Fixed config list
 - Fixed config list delimiter
+- Fixed seek shelter config option
 ### Changed:
 - Changed config from json5 to json